# faster `min` and `max`